[Docs] 백엔트 미니 팀 프로젝트 코드 반영 백엔트 미니 팀 프로젝트 코드 반영

+++ src/main/java/com/example/lonua/orders/service/OrdersService.java
@@ -1,9 +1,10 @@
 package com.example.lonua.orders.service;
 
 
-import com.example.lonua.config.BaseRes;
+import com.example.lonua.common.BaseRes;
 import com.example.lonua.orders.model.entity.Orders;
 import com.example.lonua.orders.model.entity.OrdersProduct;
+import com.example.lonua.orders.model.request.PatchUpdateOrdersReq;
 import com.example.lonua.orders.model.request.PostCreateOrdersReq;
 import com.example.lonua.orders.model.response.GetCreateOrdersRes;
 import com.example.lonua.orders.model.response.GetListOrdersRes;
@@ -24,7 +25,6 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import org.springframework.data.domain.Page;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.util.*;
@@ -39,14 +39,14 @@ public class OrdersService {
     private final OrdersProductRepository ordersProductRepository;
     private final ProductCountRepository productCountRepository;
 
-    @Transactional
+    @Transactional(readOnly = false)
     public BaseRes createOrder(User user, PostCreateOrdersReq postCreateOrdersReq) {
         Orders orders = Orders.builder()
                 .user(user)
                 .impUid(postCreateOrdersReq.getImpUid())
                 .createdAt(LocalDateTime.now().format(DateTimeFormatter.ofPattern("yyyy/MM/dd HH:mm:ss")))
                 .updatedAt(LocalDateTime.now().format(DateTimeFormatter.ofPattern("yyyy/MM/dd HH:mm:ss")))
-                .status(true)
+                .status("주문 접수")
                 .build();
 
         orders = ordersRepository.save(orders);
@@ -110,7 +110,7 @@ public class OrdersService {
     }
 
     // 주문 내역 조회
-    @Transactional
+    @Transactional(readOnly = true)
     public BaseRes list(User user, Integer page, Integer size) {
 
         Pageable pageable = PageRequest.of(page-1, size);
@@ -145,8 +145,9 @@ public class OrdersService {
     }
 
     // 주문 상세 내역 조회
-    public BaseRes read(User user, Integer idx) {
-        Optional<Orders> result = ordersRepository.findOrders(idx);
+    @Transactional(readOnly = true)
+    public BaseRes read(User user, Integer ordersIdx, Integer productIdx) {
+        Optional<Orders> result = ordersRepository.findOrders(ordersIdx, productIdx);
 
         if (result.isPresent()){
             Orders orders = result.get();
@@ -185,26 +186,60 @@ public class OrdersService {
         }
     }
 
-    @Transactional
-    public BaseRes delete(Integer idx) {
-        Integer result1 = ordersProductRepository.deleteByOrders_OrdersIdx(idx);
-        Integer result2 = ordersRepository.deleteByOrdersIdx(idx);
+    @Transactional(readOnly = false)
+    public BaseRes updateStatus(PatchUpdateOrdersReq patchUpdateOrdersReq) {
+        Optional<Orders> result = ordersRepository.findByOrdersIdx(patchUpdateOrdersReq.getOrdersIdx());
+        if(result.isPresent()) {
+            Orders orders = result.get();
+            orders.updateStatus(patchUpdateOrdersReq);
+            orders.setUpdatedAt(LocalDateTime.now().format(DateTimeFormatter.ofPattern("yyyy/MM/dd HH:mm:ss")));
+            ordersRepository.save(orders);
 
-        if(!result1.equals(0) && !result2.equals(0)) {
             return BaseRes.builder()
                     .code(200)
                     .isSuccess(true)
-                    .message("요청 성공")
-                    .result("상품이 삭제되었습니다.")
+                    .message("주문 상태 수정 성공")
+                    .result(patchUpdateOrdersReq.getStatus())
                     .build();
         } else {
             return BaseRes.builder()
                     .code(400)
                     .isSuccess(false)
-                    .message("요청 실패")
-                    .result("상품을 찾을 수 없습니다.")
+                    .message("주문 상태 수정 실패")
+                    .result("잘못된 요청입니다.")
                     .build();
         }
+    }
+    @Transactional(readOnly = false)
+    public BaseRes cancle(Integer idx) {
+        Optional<Orders> ordersStauts = ordersRepository.findByOrdersIdx(idx);
+        if(ordersStauts.isPresent()) {
+            Orders orders = ordersStauts.get();
+            if(orders.getStatus().equals("배송 전") || orders.getStatus().equals("주문 접수")) {
+                Integer result1 = ordersProductRepository.deleteByOrders_OrdersIdx(idx);
+                Integer result2 = ordersRepository.deleteByOrdersIdx(idx);
+
+                return BaseRes.builder()
+                        .code(200)
+                        .isSuccess(true)
+                        .message("요청 성공")
+                        .result("상품이 삭제되었습니다.")
+                        .build();
+            }
+        } else {
+            return BaseRes.builder()
+                    .code(400)
+                    .isSuccess(false)
+                    .message("요청 실패")
+                    .result("잘못된 요청입니다.")
+                    .build();
+        }
+            return BaseRes.builder()
+                    .code(400)
+                    .isSuccess(false)
+                    .message("요청 실패")
+                    .result("잘못된 요청입니다.")
+                    .build();
     }
